+++ assets/worldspaces/test.lua
@@ -1,0 +1,3 @@
+playerEty = getPlayerEntity()
+
+log("test-world script has been initialized; player entity has ID " .. playerEty)