+++ library/Kathu/Entity/Prototype.hs
@@ -1,27 +1,24 @@
 {-# OPTIONS_GHC -fno-warn-missing-signatures #-}
--- newFromPrototype is generated on-the-fly through TH without an explicit type signature
+-- newFromSimplePrototype is generated on-the-fly through TH without an explicit type signature
 
 {-# LANGUAGE FlexibleContexts, FlexibleInstances, MonoLocalBinds, TypeOperators, UndecidableInstances #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE TemplateHaskell #-}
 
-module Kathu.Entity.Prototype (EntityPrototype, newFromPrototype, getPrototypeID) where
+module Kathu.Entity.Prototype (EntityPrototype(..), newFromSimplePrototype, getPrototypeID) where
 
-import Data.Maybe
+import           Kathu.Entity.Components
+import           Kathu.Entity.Physics.BodyConfig    (BodyConfig)
+import           Kathu.Entity.PrototypeTemplate
+import           Kathu.Entity.SerializableComponent
+import           Kathu.Scripting.Lua.Types          (Script)
+import           Kathu.Util.Collection              (fromJustElseError)
+import           Kathu.Util.Types                   (Identifier)
 
-import Kathu.Entity.Components
-import Kathu.Entity.Physics.BodyConfig (BodyConfig, setBodyConfig)
-import Kathu.Entity.PrototypeTemplate
-import Kathu.Entity.SerializableComponents
-import Kathu.Util.Types (Identifier)
-
-defineData          "EntityPrototype"  "" ((SerializableComponent ''BodyConfig False []):serializableComponentConfigs)
-defineEntityCreator "newFromPrototypeSansPhysics" "" serializableComponentConfigs
-
--- this attaches the custom physics configuration type to our custom component's loadFromPrototype
-newFromPrototype proto = newFromPrototypeSansPhysics proto >>= \ety -> (setBodyConfig ety . bodyConfig) proto >> pure ety
+defineData          "EntityPrototype"  "" ((SerializableComponent ''BodyConfig False []):(SerializableComponent ''Script True []):serializableComponentConfigs)
+defineEntityCreator "newFromSimplePrototype" "" serializableComponentConfigs
 
 getPrototypeID :: EntityPrototype g -> Identifier
-getPrototypeID = identifier . fromMaybe (error "Attempted to load entity without Identity") . identity
+getPrototypeID = identifier . fromJustElseError "Attempted to load entity without Identity" . identity
 
-defineEntityFromJSON 'getPrototypeID ''EntityPrototype "" ((SerializableComponent ''BodyConfig False []):serializableComponentConfigs)
+defineEntityFromJSON 'getPrototypeID ''EntityPrototype "" ((SerializableComponent ''BodyConfig False []):(SerializableComponent ''Script True []):serializableComponentConfigs)